Refactored code, renamed quantity to installment, and added default value
Two major changes have occurred:  "quantity" is renamed to "installment" in Recurrence and Transaction models and additional handling done in the Transaction model to set a default value of 1 to the new "installment" column. This change impacts the data mappings and migration files. The renaming better represents the semantics of the data.