Add user authentication and modify controller structure
This commit takes significant steps for user authentication. It introduces new methods for generating and validating authentication tokens as well as getting user details by email for login functionality. In addition, the structure of controllers have been altered where BaseController functionality has been moved to GenericController. This also involves related changes in UserService and UserRepository. Several new NuGet packages for authentication and identity management are added to the solution as part of this change set.