Update category and user services logic
User service methods are updated to ensure error messages are displayed correctly when updating or resetting passwords. Extra validation is added in CategoryService to consider the AccountId while checking if a category exists. Also, in the userService UpdateAsync method, the password hash of the stored user data is assigned to the updated model to ensure correct password verification.